+++ android/KMEA/app/src/main/java/com/tavultesoft/kmea/KeyboardPickerActivity.java
@@ -18,6 +18,7 @@ import com.tavultesoft.kmea.cloud.CloudDataJsonUtil;
 import com.tavultesoft.kmea.data.CloudRepository;
 import com.tavultesoft.kmea.data.Dataset;
 import com.tavultesoft.kmea.data.Keyboard;
+import com.tavultesoft.kmea.data.KeyboardController;
 import com.tavultesoft.kmea.data.LexicalModel;
 import com.tavultesoft.kmea.util.MapCompat;
 
@@ -26,6 +27,7 @@ import androidx.appcompat.app.AlertDialog;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Typeface;
+//import android.inputmethodservice.Keyboard;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.MenuItem;
@@ -52,9 +54,7 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
 
   public static final String  KMKEY_INTERNAL_NEW_KEYBOARD = "_internal_new_keyboard_";
 
-  // TODO: Refactoring to remove keyboard selection into own keyboard manager class (MVC)
-  // Lists of installed keyboards and installed lexical models
-  private static ArrayList<HashMap<String, String>> keyboardsList = null;
+  // List of  installed lexical models
   private static ArrayList<HashMap<String, String>> lexicalModelsList = null;
   private static Dataset storageDataset = null;
 
@@ -62,7 +62,7 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
   private static int selectedIndex = 0;
   private static final String TAG = "KeyboardPickerActivity";
 
-  protected static int selectedIndex() {
+  public static int selectedIndex() {
     return selectedIndex;
   }
 
@@ -105,29 +105,6 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
 
     listView = (ListView) findViewById(R.id.listView);
 
-    //TODO: put into a logic class
-    keyboardsList = getKeyboardsList(context);
-    if (keyboardsList == null) {
-      keyboardsList = new ArrayList<HashMap<String, String>>();
-      HashMap<String, String> kbInfo = new HashMap<String, String>();
-      kbInfo.put(KMManager.KMKey_PackageID, KMManager.KMDefault_PackageID);
-      kbInfo.put(KMManager.KMKey_KeyboardID, KMManager.KMDefault_KeyboardID);
-      kbInfo.put(KMManager.KMKey_LanguageID, KMManager.KMDefault_LanguageID);
-      kbInfo.put(KMManager.KMKey_KeyboardName, KMManager.KMDefault_KeyboardName);
-      kbInfo.put(KMManager.KMKey_LanguageName, KMManager.KMDefault_LanguageName);
-      kbInfo.put(KMManager.KMKey_KeyboardVersion, KMManager.getLatestKeyboardFileVersion(
-        context, KMManager.KMDefault_PackageID, KMManager.KMDefault_KeyboardID));
-      kbInfo.put(KMManager.KMKey_Font, KMManager.KMDefault_KeyboardFont);
-      keyboardsList.add(kbInfo);
-
-      // We'd prefer not to overwrite a file if it exists
-      File file = new File(context.getDir("userdata", Context.MODE_PRIVATE),
-        KMManager.KMFilename_KeyboardsList);
-      if (!file.exists()) {
-        saveList(context, KMManager.KMFilename_KeyboardsList);
-      }
-    }
-
     lexicalModelsList = getLexicalModelsList(context);
     if (lexicalModelsList == null) {
       lexicalModelsList = new ArrayList<HashMap<String, String>>();
@@ -157,6 +134,7 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
     listView.setOnItemLongClickListener(new OnItemLongClickListener() {
       @Override
       public boolean onItemLongClick(AdapterView<?> parent, View view, final int position, long id) {
+        // Prevent the default keyboard from being removed
         if (position > 0 && canRemoveKeyboard) {
           PopupMenu popup = new PopupMenu(context, view);
           popup.getMenuInflater().inflate(R.menu.popup, popup.getMenu());
@@ -178,7 +156,7 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
       }
     });
 
-    int curKbPos = getCurrentKeyboardIndex();
+    int curKbPos = KeyboardController.getInstance().getKeyboardIndex(KMKeyboard.currentKeyboard());
     setSelection(curKbPos);
 
     KMKeyboard.addOnKeyboardEventListener(new KeyboardEventHandler.OnKeyboardEventListener() {
@@ -189,16 +167,23 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
 
       @Override
       public void onKeyboardChanged(String newKeyboard) {
-          int _index = getKeyboardIndex(context,newKeyboard);
+          int _index = KeyboardController.getInstance().getKeyboardIndex(newKeyboard);
           if(_index>=0)
           {
-            Map<String,String> _keyboard = keyboardsList.get(_index);
-            if(_keyboard==null)
+            Keyboard _keyboard = KeyboardController.getInstance().getKeyboardInfo(_index);
+            if(_keyboard==null) {
               return;
-            if(_keyboard.get(KMKEY_INTERNAL_NEW_KEYBOARD)==null)
+            }
+            if(!_keyboard.getNewKeyboard()) {
               return;
-            _keyboard.remove(KMKEY_INTERNAL_NEW_KEYBOARD);
-            saveList(context, KMManager.KMFilename_KeyboardsList);
+            }
+
+            if (_keyboard.getNewKeyboard()) {
+              _keyboard.setNewKeyboard(false);
+              // Update entry
+              KeyboardController.getInstance().add(_keyboard);
+              KeyboardController.getInstance().save(context);
+            }
             notifyKeyboardsUpdate(context);
           }
       }
@@ -223,7 +208,7 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
       adapter.notifyDataSetChanged();
     }
 
-    int curKbPos = getCurrentKeyboardIndex();
+    int curKbPos = KeyboardController.getInstance().getKeyboardIndex(KMKeyboard.currentKeyboard());
     setSelection(curKbPos);
     if (!shouldCheckKeyboardUpdates)
       return;
@@ -252,41 +237,9 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
     finish();
   }
 
-  private static int getCurrentKeyboardIndex() {
-    int pos = 0;
-
-    if (keyboardsList != null) {
-      int length = keyboardsList.size();
-      for (int i = 0; i < length; i++) {
-        HashMap<String, String> kbInfo = keyboardsList.get(i);
-        String langId = kbInfo.get(KMManager.KMKey_LanguageID);
-        String kbId = kbInfo.get(KMManager.KMKey_KeyboardID);
-        String kbKey = String.format("%s_%s", langId, kbId);
-        if (kbKey.equals(KMKeyboard.currentKeyboard())) {
-          pos = i;
-          break;
-        }
-      }
-    }
-
-    return pos;
-  }
-
-  // Previously, keyboard from package = custom keyboard
-  // Now, keyboard from package has a package ID != "cloud"
-  protected static boolean hasKeyboardFromPackage() {
-    for(HashMap<String, String> kbInfo: keyboardsList) {
-      String pkgID = MapCompat.getOrDefault(kbInfo, KMManager.KMKey_PackageID, KMManager.KMDefault_UndefinedPackageID);
-      if (!pkgID.equalsIgnoreCase(KMManager.KMDefault_UndefinedPackageID)) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
   private static boolean saveList(Context context, String listName) {
     boolean result;
+    List<Keyboard> keyboardsList = KeyboardController.getInstance().get();
     try {
       File file = new File(context.getDir("userdata", Context.MODE_PRIVATE), listName);
       ObjectOutputStream outputStream = new ObjectOutputStream(new FileOutputStream(file));
@@ -303,20 +256,6 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
       result = false;
     }
 
-    return result;
-  }
-
-  /**
-   * Save the list of installed keyboards
-   * @param context
-   * @param list
-   * @return boolean - Status if the keyboard list was successfully saved
-   */
-  protected static boolean updateKeyboardsList(Context context, ArrayList<HashMap<String, String>> list) {
-    boolean result;
-    keyboardsList = list;
-    result = saveList(context, KMManager.KMFilename_KeyboardsList);
-    notifyKeyboardsUpdate(context);
     return result;
   }
 
@@ -347,65 +286,35 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
    */
   private static void switchKeyboard(int position, boolean aPrepareOnly) {
     setSelection(position);
-    int listPosition = (position >= keyboardsList.size()) ? keyboardsList.size()-1 : position;
-    HashMap<String, String> kbInfo = keyboardsList.get(listPosition);
-    String pkgId = kbInfo.get(KMManager.KMKey_PackageID);
-    if (pkgId == null || pkgId.isEmpty()) {
-      pkgId = KMManager.KMDefault_UndefinedPackageID;
-    }
-    String kbId = kbInfo.get(KMManager.KMKey_KeyboardID);
-    String langId = kbInfo.get(KMManager.KMKey_LanguageID);
-    String kbName = kbInfo.get(KMManager.KMKey_KeyboardName);
-    String langName = kbInfo.get(KMManager.KMKey_LanguageName);
-    String kFont = kbInfo.get(KMManager.KMKey_Font);
-    String kOskFont = kbInfo.get(KMManager.KMKey_OskFont);
+    int size = KeyboardController.getInstance().get().size();
+    int listPosition = (position >= size) ? size-1 : position;
+    Keyboard kbInfo = KeyboardController.getInstance().getKeyboardInfo(listPosition);
+    String pkgId = kbInfo.getPackageID();
+    String kbId = kbInfo.getKeyboardID();
+    String langId = kbInfo.getLanguageID();
+    String kbName = kbInfo.getKeyboardName();
     if(aPrepareOnly)
       KMManager.prepareKeyboardSwitch(pkgId, kbId, langId, kbName);
     else
-      KMManager.setKeyboard(pkgId, kbId, langId, kbName, langName, kFont, kOskFont);
+      KMManager.setKeyboard(kbInfo);
   }
 
-  protected static boolean addKeyboard(Context context, HashMap<String, String> keyboardInfo) {
+  protected static boolean addKeyboard(Context context, Keyboard keyboardInfo) {
     boolean result = false;
-
-    if (keyboardsList == null) {
-      keyboardsList = getKeyboardsList(context);
-    }
-
-    if (keyboardsList == null) {
-      keyboardsList = new ArrayList<HashMap<String, String>>();
-    }
+    List<Keyboard> keyboardsList = KeyboardController.getInstance().get();
 
     if (keyboardInfo != null) {
-      String pkgID = keyboardInfo.get(KMManager.KMKey_PackageID);
-      String kbID = keyboardInfo.get(KMManager.KMKey_KeyboardID);
-      String langID = keyboardInfo.get(KMManager.KMKey_LanguageID);
-
       // TODO:  Possible optimization - do we have anything with this language code already?
       //        Only invalidate the lexical cache if not.
       CloudRepository.shared.invalidateLexicalModelCache(context);
 
-      if (pkgID != null && kbID != null && langID != null) {
-        String kbKey = String.format("%s_%s", langID, kbID);
-        if (kbKey.length() >= 3) {
-          int x = getKeyboardIndex(context, kbKey);
-          if (x >= 0) {
-            if(keyboardsList.get(x).get(KMKEY_INTERNAL_NEW_KEYBOARD)!=null)
-              keyboardInfo.put(KMKEY_INTERNAL_NEW_KEYBOARD,KMKEY_INTERNAL_NEW_KEYBOARD);
-            keyboardsList.set(x, keyboardInfo);
-            result = saveList(context, KMManager.KMFilename_KeyboardsList);
-          } else {
-            keyboardInfo.put(KMKEY_INTERNAL_NEW_KEYBOARD,KMKEY_INTERNAL_NEW_KEYBOARD);
-            keyboardsList.add(keyboardInfo);
-            result = saveList(context, KMManager.KMFilename_KeyboardsList);
-            if (!result) {
-              keyboardsList.remove(keyboardsList.size() - 1);
-            }
-          }
-        }
+      keyboardInfo.setNewKeyboard(true);
+      KeyboardController.getInstance().add(keyboardInfo);
+      result = KeyboardController.getInstance().save(context);
+      if (!result) {
+        Log.e(TAG, "addKeyboard failed to save");
       }
     }
-
     notifyKeyboardsUpdate(context);
 
     return result;
@@ -453,13 +362,11 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
 
   protected static boolean removeKeyboard(Context context, int position) {
     boolean result = false;
-    if (keyboardsList == null) {
-      keyboardsList = getKeyboardsList(context);
-    }
 
-    if (keyboardsList != null && position >= 0 && position < keyboardsList.size()) {
-      keyboardsList.remove(position);
-      result = saveList(context, KMManager.KMFilename_KeyboardsList);
+    // Prevent the first keyboard (index 0) from being removed
+    if (position > 0) {
+      KeyboardController.getInstance().remove(position);
+      result = KeyboardController.getInstance().save(context);
     }
 
     notifyKeyboardsUpdate(context);
@@ -468,7 +375,7 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
   }
 
   protected static void deleteKeyboard(Context context, int position) {
-    int curKbPos = getCurrentKeyboardIndex();
+    int curKbPos = KeyboardController.getInstance().getKeyboardIndex(KMKeyboard.currentKeyboard());
     boolean result = removeKeyboard(context, position);
 
     if (result) {
@@ -480,7 +387,7 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
       if (position == curKbPos && listView != null) {
         switchKeyboard(0,false);
       } else if(listView != null) { // A bit of a hack, since LanguageSettingsActivity calls this method too.
-        curKbPos = getCurrentKeyboardIndex();
+        curKbPos = KeyboardController.getInstance().getKeyboardIndex(KMKeyboard.currentKeyboard());
         setSelection(curKbPos);
       }
     }
@@ -577,31 +484,7 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
       return storageDataset;
     }
 
-    List<HashMap<String, String>> kbdMapList = getKeyboardsList(context);
-    if (kbdMapList == null) {
-      kbdMapList = new ArrayList<>(0);
-    }
-    List<Keyboard> kbdsList = new ArrayList<>(kbdMapList.size());
-
-    for(HashMap<String, String> kbdMap: kbdMapList) {
-      boolean isNewKeyboard = kbdMap.containsKey(KeyboardPickerActivity.KMKEY_INTERNAL_NEW_KEYBOARD) &&
-        kbdMap.get(KeyboardPickerActivity.KMKEY_INTERNAL_NEW_KEYBOARD).equals(KeyboardPickerActivity.KMKEY_INTERNAL_NEW_KEYBOARD);
-
-      Keyboard k = new Keyboard(
-        kbdMap.get(KMManager.KMKey_PackageID),
-        kbdMap.get(KMManager.KMKey_KeyboardID),
-        kbdMap.get(KMManager.KMKey_KeyboardName),
-        kbdMap.get(KMManager.KMKey_LanguageID),
-        kbdMap.get(KMManager.KMKey_LanguageName),
-        MapCompat.getOrDefault(kbdMap, KMManager.KMKey_Version, "1.0"),
-        MapCompat.getOrDefault(kbdMap, KMManager.KMKey_CustomHelpLink, ""),
-        isNewKeyboard,
-        MapCompat.getOrDefault(kbdMap, KMManager.KMKey_Font, null),
-        MapCompat.getOrDefault(kbdMap, KMManager.KMKey_OskFont, null)
-      );
-      kbdsList.add(k);
-    }
-
+    List<Keyboard> kbdsList = getKeyboardsList(context);
     List<HashMap<String, String>> lexMapList = getLexicalModelsList(context);
     if(lexMapList == null) {
       lexMapList = new ArrayList<>(0);
@@ -636,26 +519,7 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
     storage.keyboards.setNotifyOnChange(false);
     storage.keyboards.clear();
 
-    List<HashMap<String, String>> mapList = getKeyboardsList(context);
-    List<Keyboard> kbdList = new ArrayList<>(mapList.size());
-    for(HashMap<String, String> kbdMap: mapList) {
-      boolean isNewKeyboard = kbdMap.containsKey(KeyboardPickerActivity.KMKEY_INTERNAL_NEW_KEYBOARD) &&
-        kbdMap.get(KeyboardPickerActivity.KMKEY_INTERNAL_NEW_KEYBOARD).equals(KeyboardPickerActivity.KMKEY_INTERNAL_NEW_KEYBOARD);
-
-      Keyboard k = new Keyboard(
-        kbdMap.get(KMManager.KMKey_PackageID),
-        kbdMap.get(KMManager.KMKey_KeyboardID),
-        kbdMap.get(KMManager.KMKey_KeyboardName),
-        kbdMap.get(KMManager.KMKey_LanguageID),
-        kbdMap.get(KMManager.KMKey_LanguageName),
-        MapCompat.getOrDefault(kbdMap, KMManager.KMKey_Version, "1.0"),
-        MapCompat.getOrDefault(kbdMap, KMManager.KMKey_CustomHelpLink, ""),
-        isNewKeyboard,
-        MapCompat.getOrDefault(kbdMap, KMManager.KMKey_Font, null),
-        MapCompat.getOrDefault(kbdMap, KMManager.KMKey_OskFont, null)
-      );
-      kbdList.add(k);
-    }
+    List<Keyboard> kbdList = KeyboardController.getInstance().get();
     storage.keyboards.addAll(kbdList);
     storage.keyboards.notifyDataSetChanged();
   }
@@ -684,31 +548,12 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
     storage.lexicalModels.notifyDataSetChanged();
   }
 
-  protected static ArrayList<HashMap<String, String>> getKeyboardsList(Context context) {
-    return getList(context, KMManager.KMFilename_KeyboardsList);
+  protected static List<Keyboard> getKeyboardsList(Context context) {
+    return KeyboardController.getInstance().get();
   }
 
   protected static ArrayList<HashMap<String, String>> getLexicalModelsList(Context context) {
     return getList(context, KMManager.KMFilename_LexicalModelsList);
-  }
-
-  protected static boolean containsKeyboard(Context context, String keyboardKey) {
-    if (keyboardsList == null) {
-      keyboardsList = getKeyboardsList(context);
-    }
-
-    String kbKeys = "";
-    if (keyboardsList != null) {
-      int length = keyboardsList.size();
-      for (int i = 0; i < length; i++) {
-        kbKeys += String.format("%s_%s", keyboardsList.get(i).get(KMManager.KMKey_LanguageID), keyboardsList.get(i).get(KMManager.KMKey_KeyboardID));
-        if (i < length - 1) {
-          kbKeys += ",";
-        }
-      }
-    }
-
-    return kbKeys.contains(keyboardKey);
   }
 
   public static boolean containsLexicalModel(Context context, String lexicalModelKey) {
@@ -729,84 +574,6 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
     }
 
     return false;
-  }
-
-  protected static int getCurrentKeyboardIndex(Context context) {
-    int index = -1;
-
-    if (keyboardsList == null) {
-      keyboardsList = getKeyboardsList(context);
-    }
-
-    if (keyboardsList != null) {
-      int length = keyboardsList.size();
-      for (int i = 0; i < length; i++) {
-        HashMap<String, String> kbInfo = keyboardsList.get(i);
-        String langId = kbInfo.get(KMManager.KMKey_LanguageID);
-        String kbId = kbInfo.get(KMManager.KMKey_KeyboardID);
-        String kbKey = String.format("%s_%s", langId, kbId);
-        if (kbKey.equals(KMKeyboard.currentKeyboard())) {
-          index = i;
-          break;
-        }
-      }
-    }
-
-    return index;
-  }
-
-  protected static HashMap<String, String> getCurrentKeyboardInfo(Context context) {
-    int index = getCurrentKeyboardIndex(context);
-    if (index >= 0) {
-      return keyboardsList.get(index);
-    } else {
-      return null;
-    }
-  }
-
-  protected static int getKeyboardIndex(Context context, String keyboardKey) {
-    int index = -1;
-
-    if (keyboardsList == null) {
-      keyboardsList = getKeyboardsList(context);
-    }
-
-    if (keyboardsList != null) {
-      int length = keyboardsList.size();
-      for (int i = 0; i < length; i++) {
-        HashMap<String, String> kbInfo = keyboardsList.get(i);
-        String langId = kbInfo.get(KMManager.KMKey_LanguageID);
-        String kbId = kbInfo.get(KMManager.KMKey_KeyboardID);
-        String kbKey = String.format("%s_%s", langId, kbId);
-        if (kbKey.equals(keyboardKey)) {
-          index = i;
-          break;
-        }
-      }
-    }
-
-    return index;
-  }
-
-  /**
-   * Get the list of associated keyboards for a given language ID
-   * @param langId
-   * @return ArrayList of keyboard
-   */
-  public static ArrayList<HashMap<String, String>> getAssociatedKeyboards(String langId) {
-    if (keyboardsList != null) {
-      ArrayList<HashMap<String, String>> associatedKeyboards = new ArrayList<HashMap<String, String>>();
-      for (HashMap<String, String> keyboardInfo: keyboardsList) {
-        if (keyboardInfo.get(KMManager.KMKey_LanguageID).equalsIgnoreCase(langId)) {
-          keyboardInfo.put(KMManager.KMKey_Icon, String.valueOf(R.drawable.ic_arrow_forward));
-          keyboardInfo.put("isEnabled", "true");
-          associatedKeyboards.add(keyboardInfo);
-        }
-      }
-      return associatedKeyboards;
-    }
-
-    return null;
   }
 
   /**
@@ -840,27 +607,6 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
     return index;
   }
 
-  protected static HashMap<String, String> getKeyboardInfo(Context context, int index) {
-    if (index < 0) {
-      return null;
-    }
-
-    if (keyboardsList == null) {
-      keyboardsList = getKeyboardsList(context);
-    }
-
-    if (keyboardsList != null && index < keyboardsList.size()) {
-      HashMap<String, String> kbInfo = keyboardsList.get(index);
-      String pkgID = kbInfo.get(KMManager.KMKey_PackageID);
-      if (pkgID == null || pkgID.isEmpty()) {
-        kbInfo.put(KMManager.KMKey_PackageID, KMManager.KMDefault_UndefinedPackageID);
-      }
-      return kbInfo;
-    }
-
-    return null;
-  }
-
   protected static HashMap<String, String> getLexicalModelInfo(Context context,int index) {
     if (index < 0) {
       return null;
@@ -884,16 +630,19 @@ public final class KeyboardPickerActivity extends AppCompatActivity {
   }
 
   public static void handleDownloadedKeyboard(Context context, HashMap<String, String> keyboardInfo) {
-    String keyboardID = keyboardInfo.get(KMManager.KMKey_KeyboardID);
-    String languageID = keyboardInfo.get(KMManager.KMKey_LanguageID);
-    String kbKey = String.format("%s_%s", languageID, keyboardID);
-    int index = getKeyboardIndex(context, kbKey);
-    if (index == -1) {
-      // Add the downloaded keyboard if not found
-      addKeyboard(context, keyboardInfo);
-      index = getKeyboardIndex(context, kbKey);
-    }
-    keyboardsList.set(index, keyboardInfo);
-    saveList(context, KMManager.KMFilename_KeyboardsList);
+    boolean isNewKeyboard = true;
+    Keyboard k = new Keyboard(
+      keyboardInfo.get(KMManager.KMKey_PackageID),
+      keyboardInfo.get(KMManager.KMKey_KeyboardID),
+      keyboardInfo.get(KMManager.KMKey_KeyboardName),
+      keyboardInfo.get(KMManager.KMKey_LanguageID),
+      keyboardInfo.get(KMManager.KMKey_LanguageName),
+      keyboardInfo.get(KMManager.KMKey_KeyboardVersion),
+      keyboardInfo.get(KMManager.KMKey_HelpLink),
+      isNewKeyboard,
+      keyboardInfo.get(KMManager.KMKey_Font),
+      keyboardInfo.get(KMManager.KMKey_OskFont));
+    KeyboardController.getInstance().add(k);
+    KeyboardController.getInstance().save(context);
   }
 }