+++ web/tools/recorder/build.sh
@@ -7,10 +7,11 @@
 ## START STANDARD BUILD SCRIPT INCLUDE
 # adjust relative paths as necessary
 THIS_SCRIPT="$(greadlink -f "${BASH_SOURCE[0]}" 2>/dev/null || readlink -f "${BASH_SOURCE[0]}")"
-SCRIPT_DIR="$(dirname "$THIS_SCRIPT")"
 . "$(dirname "$THIS_SCRIPT")/../../../resources/build/build-utils.sh"
 . "$KEYMAN_ROOT/resources/shellHelperFunctions.sh"
 ## END STANDARD BUILD SCRIPT INCLUDE
+
+SCRIPT_DIR="$(dirname "$THIS_SCRIPT")"
 
 # Ensure the dependencies are downloaded.  --no-optional should help block fsevents warnings.
 verify_npm_setup
@@ -36,7 +37,7 @@ compilecmd="$compiler"
 
 $compilecmd -p "$SCRIPT_DIR/tsconfig.json"
 if [ $? -ne 0 ]; then
-    fail "KeymanWeb test sequence recorder compilation failed; aborting full Recorder compilation."
+    fail "KeymanWeb test sequence recorder compilation failed."
 fi
 
 PRODUCT="$SCRIPT_DIR/$OUTPUT_DIR/$OUTPUT"